Add calico service nrpe check.
